Refactored Domain Model
since Module is completly unnecessary it has been removed. All
Logbookentries are now directly assigned to a project and/or a issue.
Assigning a logbookentry to an issue requires the issue to be assigned
to a logbookentry in order to transitivly add the entry to the project
aswell. But a LogbookEntry can also only be assigned to a project
itself, without beeing added to an issue